requirements_doc.txt
sphinx >= 1.3.1
sphinx_rtd_theme >= 0.1.8
Pygments >= 2.0.1
numpydoc >= 0.5
